#18 <feat> <(Review)> : Dto 생성 및 Service, Repository class, interface 생성

+++ src/main/java/com/example/skillback/common/domain/review/dto/ReviewListResponse.java
@@ -12,7 +12,7 @@ import lombok.NoArgsConstructor;
 public class ReviewListResponse {
 
     private String reviewWriter;
-    private Long reviewScore;
+    private int reviewScore;
     private String reviewContent;
     private Long likeCnt;
     private Long hateCnt;
